hw0 minor edits
UserMealWithExcess:
Rewrote equals() and hashcode().
Removed the constructor that uses UserMeal, and in the UserMealsUtil class created a method that converts UserMeal to UserMealExcess.
Updated and clarified javadoc in methods.

Made line wrapping refactoring in UserMealWithExcess and UserMealsUtilTest classes.
Moved the declaration of the result variable closer to the place of use.
Replaced LinkedList with ArrayList.